+++ CHANGELOG.md
@@ -13,3 +13,4 @@ Release summary...
 
 - A user can no longer enter an invalid C++ variable name.
 - XRC files now include a comment warning that regeneration will eliminate any hand edits.
+- `Generate XRC files...` on Tools menu now launches a dialog allowing you to choose between combined or individually generated XRC files.